Differentiated between austin, versions. .gitignore and README updated
